+++ src/function/lsp/completion/vlog.ts
@@ -108,7 +108,7 @@ class VlogMacroCompletionProvider implements vscode.CompletionItemProvider {
 
 class VlogPositionPortProvider implements vscode.CompletionItemProvider {
     public async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext): Promise<vscode.CompletionItem[] | vscode.CompletionList<vscode.CompletionItem> | null | undefined> {
-        console.log('enter VlogPositionPortProvider');
+        // console.log('enter VlogPositionPortProvider');
         
         try {
             const suggestPositionPorts: vscode.CompletionItem[] = [];