+++ src/test/manager/user/src/hello.v
@@ -1,6 +1,0 @@
-module hello(
-    input clk, reset,
-    output value
-);
-
-endmodule