+++ src/test/monitor/src2/add.v
@@ -1,9 +1,0 @@
-module SimpleAdd_2(
-    input [7:0] a, b,
-    output [7:0] c
-);
-
-    assign c = a + b;
-    
-    
-endmodule //SimpleAdd