+++ src/test/vlog/dependence_test/child_2.v
@@ -1,8 +1,0 @@
-module dependence_2 (
-    input a, b, c,
-    output Q 
-);
-
-  assign Q = a & b | ((b & c) & (b | c));
-
-endmodule